% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_app.R
\name{run_app}
\alias{run_app}
\title{Shiny App Bio stats}
\usage{
run_app(...)
}
\arguments{
\item{...}{Arguments passed to [shiny::runApp()], such as `port`, `host`,
and `launch.browser`.}
}
\description{
Run this function to start the shiny app in the browser.
}
